pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.1",
    "celery>=5.5.3",
    "fastapi[standard]>=0.121.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "redisvl>=0.12.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
